整理了客户端代码，以方便做ilruntime兼容

+++ Unity/Assets/Editor/BehaviorTreeEditor/Event/BehaviorTreeReplaceClickEvent_ReplaceNode.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 namespace MyEditor
 {
-	[Event(EventIdType.BehaviorTreeReplaceClick)]
+	[Event((int)EventIdType.BehaviorTreeReplaceClick)]
 	public class BehaviorTreeReplaceClickEvent_ReplaceNode: IEvent<string, Vector2>
 	{
 		public void Run(string name, Vector2 pos)
